NestedQuestions should use Card convenience method
Remember that currently the NestedQuestion controller actually returns
CardContent and not NestedQuestions.  Therefore, it would be premature
for the Ember front end to know about CardVersions when working with
NestedQuestions.

So in the meantime, look up CardContent via the Card and use the new
Card convenience method to do it.