Assign Card to all Tasks
All Task answerables should have a Card assigned via a CardVersion, so
ensure that Tasks created through the front end controller (admin) are
given CardVersions and be sure that all tests which require Cards to
be loaded for that Task is loaded before running the test.

For now, we are being specific with which Cards need preloaded, so
intentionally we are *not* dropping a hammer and loading all Cards.
Instead, we only load the specific ones where they are needed.  In many
cases, this can be done when creating the Task by using the
:with_loaded_card trait.  In other cases, the CardLoader can be called.