Reassign correct CardVersion relationship for FrontMatterReviewerReports
Without running this rake task, FrontMatterReviewerReportTasks will be
incorrectly associated to ReviewerReport Cards.  This means that the
wrong set of questions will be tied to ReviewerReports in the case of
front matter.

Here's why ...

A ReviewerReport will always have a Card associated to it through the
CardVersion association.  This association helps determine the set of
questions that is displayed to the end user.  When the ReviewerReport is
initially created (using the ReviewerReportCreator service class), it
will assign one of two different types of Cards -- either a
ReviewerReport or a TahiStandardTasks::FrontMatterReviewerReport.  The
ability to determine which of these two is assigned requires intimate
knowledge of the Task instance that the Report is associated with and
cannot be done simply by class name.  Compounding this issue is the fact
that although there is an actual ReviewerReport class, there is no such
thing as a TahiStandardTasks::FrontMatterReviewerReport model.

This particular one time rake task will ensure that all
FrontMatterReviewerReportTasks will have its front matter reviewer
reports associated to the correct Card, so that the correct questions
will be shown to the user on the front end.