Remove Task#participants= setter
A setter method with side effects can lead to many headaches.  Anytime
a developer created a new Task like this:

```
Task.new(participants: [User.first])
```

the task would have actually been created (saved).  A developer would
normally expect a new (unsaved) instance to be instantiated, not created.

This caused all sorts of headaches with the TaskFactory which did not
anticipate these side effects and validations (especially database
validations) would cause it to explode.

After investigation, it appears that the only place an array of
participants was being passed in was when running specs, so this logic
was removed from Task and added to our rspec FactoryGirl factory instead.