Add (protected) Card model destroys
Cards can only be destroyed if they are in draft status.  Currently, a
state of "draft" means that the Card has only one CardVersion and it
has never been "published".  In this case, it is safe to destroy the
Card because it is not (and has never) actually been used anywhere.