Always use the latest published card version
The following PR was merged into master while the majority of work was
done with the current PR:

https://github.com/Tahi-project/tahi/pull/3029

The above PR introduces the concept of a Card being "published" or
"draft".  In almost every case throughout the system, we want to use
the latest PUBLISHED card, since a draft card is not considered finished
and should not be the one used when creating new Tasks.