Ensure all system CardVersions are published
Since the concept of published/draft Cards were introducted while this
work was being completed, move the migration to *after* the published
column has been added to the CardVersion.  Ensure that any of the
new system CardVersions that are created are also marked as published.

Otherwise, an error will be thrown when attempting to create any new
non-CustomCardTask models.