APERTA-9087 Fix problem with stale connections in pool
The `Sidekiq.redis` method depends on catching an error to remove connections
from the connection pool. We were catching the exception and not re-raising
inside a block. This meant that stale connections were not being removed from
the pool, and the write was not retried.

Move the scope of the begin/rescue block to ensure that stale connections are
removed from the pool.