Record history entry and published_by for card versions
Makes card.publish! take a user and a history entry as parameters.  Since both
are inseparable from the act of publishing I felt like it was more natural to
include them directly rather than expecting the client (meaning the controller)
to remember to set the values manually beforehand.

This commit also adds the card history information to the card version
serializer, and sends down card version ids with a given card