Move this migration after Card#state column addition
This migration triggers a data migration that creates new cards.  Cards
now require a state column to be present.  Due to the way in which 
parallel work was being done, the migration to add the state column
showed up *AFTER* this migration, causing this one to fail against
production data.