Consolidate ReviewerReport Card Lookup
When a new ReviewerReport is being created, it needs to look at its
paper to determine if the Card associated should either be a
`ReviewerReport` or a `TahiStandardTasks::FrontMatterReviewerReport`.

Prior to this PR, this same logic was being done identically in two
different places, the CreateReviewerReports service class and the 
ReviewerReportTaskCreator subscription service class.

This PR consolidates this lookup into one place, the ReviewerReport
model.

However, by doing this is became apparent that when a new Answerable
model is created, it needs to know more than just its class name to find
the correct Card model.  This means that the class method doing this
Card lookup using class name of the Answerable model needed to go away.

It would be too confusing to have a class based method that did this
and say, "well, it is *ALMOST* always accurate" because it would not
be accurate in the case of a ReviewerReport (which needs to look at its
paper) and a CustomCardTask (which needs to have a Card association 
explicitly set).