Reorder db migrations to allow data migration
The data migration for assigning cards to legacy tasks calls the
CardLoader, which eventually will try to publish any cards that are
still in a draft state via the `card.publish!` method.  The code for
publish has validations around `history_entry`, which didn't exist in
the DB when the data migration was set to run.  There was a similar
problem for `published_by`.  This commit moves the migrations for adding
the history_entry and published_by columns before the data migration.

Tested on a dump of staging taken from big hector today 5/18/2017