Provide contexts for templates
A context will wrap an object and provided a whitelisted version to
access the internal properties. This in turn gets fed into the parsed
templates for expansion.

The contexts are used as wrappers around objects and as specific
collections of objects. The latter are fed into the rendering engines.

APERTA-9857