Create template for decisions
An example template for deccisions was made for the 'RA Major
Revision' template.

Along the way, several contexts for system models were added

- PaperContext
- ReviewerReportContext
- UserContext
- RegisterDecisionContext

The first three are generic cases that map 1-1 with models. The last,
RegisterDecisionContext, maps to a set of template types. In this
case, the register decision task has templates that pull from the same
sort of data.

APERTA-9857