Add `workflow_display_only` to CardVersion.
Since every Task is now associated to a Card, we will use the CardVersion as a place to configure whether a Card's Task should be hidden from the
accordion sidebar or not on the paper dashboard page.

In practice this only applies to CustomCardTasks since they are the only
ones that currently would be configured to be hidden or not.