APERTA-9473 Fix has_many :through support for authorizations
Using the authorizations config to authorize via a has_many :through relation
where the through class belongs_to, rather than has_one, the target class,
caused the wrong join query to be created.

Fix this by checking whether the relation on the target is a has_one or
belongs_to and adjust the query accordingly.