APERTA-9951 seed registered settings
The seeding registered settings should happen a the same times in the
application lifecycle that Cards are loaded, etc.  For now it's a simple rake
task that is intended to be idempotent.

Note that RegisteredSettings are not seeded in tests.  Since Settings are first
looked up directly in Configurable we can use FactoryGirl to create the settings
we need for a given test rather than having to through the RegisteredSettings
table for every setting-related test.