APERTA-9951 associate task templates to tasks for new papers
Some tasks will have different settings per worfklow.  We decided that the
TaskTemplate for a given task is the natural place for these kinds of settings
to live, and given that a Task will need to have a reference to its
TaskTemplate.  If a Task is added to paper after the paper has been created the
link to its TaskTemplate will be nil, and we're okay with that.  The discussion
in https://jira.plos.org/jira/browse/APERTA-9495 has some details.