Created External Correspondence is sorted by date.
Attachments which follow the external correspondence is created in a
derived class of Attachmnts model. New models, serializers, and adaptera
have been made for the external-correspondece on rails and ember.

New endpoints are

 - POST /api/external_correspondence
 - POST /api/external_correspondence_attachment

Tests are yet to pass. For some reason, ember-cli complains about a
Brocfile in an engine. I am yet to figure out a fix.

Ticket: APERTA-9047