Added recipients to client correspondence model
This helps with serializing. Idealy, transforming `.recipients` from the
rails side to `.recipient` on the Ember side does not take into
consideration all the use cases of the correspondence epic. However,
since I am not responsible for this block of code, I decided not to
destroy it. Consider it deprecated.