APERTA-9473 Fix Permission.ensure_exists for states
Previously, we were not doing an exact match on the many-to-many relation
between permissions and states. This meant that `Permission.ensure_exists` could
return the wrong permission for a set of states passed in.

We now do an exact match on the array of states joined to the the permission.