Change Scenario implementation
First, change the name of collections of objects into Scenarios. This
leaves the term Context for only a single object.

To address questions of performance, the top-level Scenario objects
move from accessors to methods. This means they won't be populated
until required by Liquid.

Additionally a missing test for the `ReviewerReportScenario` was
added.

APERTA-9857