APERTA-9930 use tinymce on manuscript-new page
- Adds a 'wormhole' option on `overlay-fullscreen` which defaults to 'true'.
We have to disable ember-wormhole for this use of tinymce.  Our tinymce wrapper
component will run `tinymce.init` in its `didInsertElement` hook, which runs
before `ember-wormhole`'s
[_appendToDestination](https://github.com/yapplabs/ember-wormhole/blob/master/addon/components/ember-wormhole.js#L63)
. Ember wormhole will insert content and then move that content to a new place
in the DOM.  At that point something goes awry with tinymce.

Other instances of tinymce inside ember-wormhole are also wrapped in a
`task-load` component, which defers rendering its content until a particular
task's resources have been fetched from the server.  So far that's always taken
longer than it takes for the wormhole append, so we've never had to deal with
this phenomenon before.

In the case of manuscript-new, rendering it inline (without using the wormhole)
doesn't appear to be a problem from a UI point of view - the styles apply
correctly and it behaves as expected for one of our full screen overlays.

We may want to evaluate using https://github.com/ef4/ember-elsewhere as a replacement
for ember-wormhole for cases like these