Fix unsync'd card.rng changes
At around the same time, two commits (by two different authors) were 
introduced into `master` to allow multiple validations to be added to
the XML.  Either due to oversight of the second committer and/or merge
conflict, the `card.rng` was out of sync with the `card.rnc`.

The removal of these lines should be safe.  This removes the abstracted
definition of a `validation` xml node because apparently they are being
defined in-line where needed.