APERTA-9047 client-side validations
Since the time and the date fields are separate in the UI and combined
in the API, this forces validations on both fields to be done on the
client side.

Having the validations for the date and time fields on the UI side and
the rest of the presence validations on the API breaks UX; specifically
the case when the client validations are not met but also, the API
validations are not met. Since the API validations are never triggered,
only the time and date fields have error icons.

The choice is thus between full client-side validation and a pipeline
case of serialization, send to the server, validate, deserialize,
distinguish error message when back to the client to deal with the
separate date and time field.

The server call is expensive. Thus client-side validations were chosen.
Since these are only presence validations, simple JS logic did suffice.