Wrap academic editor invite body in <p> tags
Both the reviewer and academic editor invitation templates are
hard coded in their respective invitation models. This issue affected
both templates, but the reviewer template was being converted over to
liquid and the work to change it was done there coincidentally.  This
template hasn't been converted to liquid yet and it was also missed in
the initial triage.