APERTA-10433 we now send unsent unbilled reports
Before now, the behaviour behind sending reports was based on the date
they were accepted. This worked until FTP became unreliable. If FTP is
down for more than a day, some reports are not sent.

In order to work around this, the Activity model is employed to identify
all previously sent reports. With this, we can now send all previously
unsent reports.

I have kept the test and code which sends the reports created within the
previous 24 hours per adventure this behaviour is desired sometime in
the future.