Add explicit signature for rich-text-editor component
It was not immediately obvious exactly what properties should be sent 
to the rich-text-editor component.  This introduces propTypes which 
makes its signature easier to understand for future developers.

Set defaults of the rich-text-editor appropriately.

Always add a `data-editor` html5 data attribute for every
rich-text-editor parent div.  Previously, it was only being set if an
ident was being sent.

Remove any unused parameters when rich-text-editor was being invoked.

If the rich-text-editor is flagged as 'disabled', surround the non-
html view with a "findable" div.