Seed legacy Cards for specs
When using the FactoryGirl Paper factory and specifying `:with_tasks`
trait, it expects that legacy cards are already seeded.  It is possible
that non-legacy (custom cards) are loaded, so the query here needs to
be a bit more precise to ensure they are loaded.

We should certainly look in the future to remove the need for this trait
all together since it is rarely needed for most tests.