Default `filter_by_card_id` to nil on permission helpers
We currently have specs that use the `permission_spec_helper` to create
new Permissions or ensure that an existing Permission exists.  Because
`filter_by_card_id` was not being defaulted to `nil` as part of this
check, the `where` clause for a Permission would be too broad and would
end up finding an unintended record.

This problem was uncovered in the `custom_card_task_spec.rb` where a 
generic "you have the ability to view all tasks" permission was being
created.  However, as part of the test setup, a new Journal was being
created that created a new Cover Letter CustomCardTask which has a 
view permission for that particular card id.  Since the scope was too
broad, this permission (the one with a `filter_by_card_id` that was not
nil) was being found instead of creating one that had a nil 
`filter_by_card_id`.