Do not automatically create a Journal when defining Role
Previously, if you using the `role` spec helper function, it would
always create a new Journal to attach the role to.  This also means
that the Journal would include database records for MMTs, PhaseTemplates,
and TaskTemplates.

These changes allow you to pass in an optional Journal parameter to
create the Role for a particular Journal.