Fix flaky similarity check spec
This spec was flaky because it would actually generate a signed URL for S3, and
then stub it out, but these URLs change every second (they are signed with a
time-limit), so if a second passed between when the URL was stubbed out and when
the code was called, the test would fail.

This is fixed by stubbing out the URL that is returned by the file object to
ensure that it does not change.