APERTA-9692 can now schedule events off templates
Now there's the `ScheduledEvents` and there's the
`ScheduledEventsTemplate`. These two can be used to "automagically"
schedule events on anything which has a `DueDatetime`. But for now, this
magic would need an assistant.

Currently, these `ScheduledEventsTemplate`s should be created directly
into the database and the owner should be the name of the class which
has the DueDatetime as a string. With a little modification to any class
which has a DueDatetime, the ScheduledEventsFactory can pick these
templates up and schedule events accordingly.

Tests are yet to be written.