APERTA-9692 some review changes
Reviewing this ticket requires some console work. This ticket fixes up
the scheduled event templates created for ReviewerReports. This also
cleans up the Ember models up a bit since, for now, the reports need not
be embedded in the events. And, this commit adds the feature flag to be
included during bin/setup