Fix Competing Interests instruction statement
I left out a trailing quote and final sentence in the text shown when
choosing “yes” on the converted Competing Interest card. This fixes the
text of that in the affected CardContent, as well as cleans up the dev
database, which was showing “Custom Card” as the title of all the
converted Competing Interest cards.

APERTA-10097