APERTA-9692 in-memory templates, decoupling
The whiteboarding session to design the infrastructure being setup in
this ticket [for subsequent tickets] took into consideration templates
which the scheduled events would be created against. I originally
interpreted this as active-record models trying to anticipate the
engineering of subsequent work built on this. After some consideration
with Jack and Shawn, we re-concluded that these templates can live in
the model since they do not change much in the scope of this ticket. The
ScheduledEventsTemplate as an ActiveRecord model is no longer needed.

Originally, the `{{scheduled-events}}` component was too coupled to the
ReviewerReport. This commit also decouples these so the component can be
used with anything which has scheduled events.