Make short-input validations behave like paragraph
There was a minor difference in behavior between short-input and
paragraph-input regarding what happens when the user clicks into a
blank, but error, field. TinyMCE fires a change event when the user
focuses, so paragraph-input got a free valueChanged call. This adds
a similar check when first focusing on a short-input field.

APERTA-10338