APERTA-9687 Scheduled Event States
For the scope of this ticket, the states needed for the ScheduledEvent
are `complete`, `active`, and `inactive`.

When an event is scheduled, it is most likely that we want this event to
be triggered at some point. No event is scheduled in a deactivated
state. So `active` becomes the initial state.

An active event can be `trigger`ed. After the event id triggered, it is
assumed that the event is complete. This is only for the scope of this
ticket. There may be error states in processing. There may also be a
processing transient state, if this is to be elaborate.

Rescheduling is the focus of this ticket. Depending on the direction of
the due date, the event may transition from an active state to an
inactive state, and vice versa. A completed event may also transition to
an active (yet to be triggered) event. But for history purposes, this
event would be duplicated as an active one.