APERTA-9687 Rescheduling already existing events
The one thing which does not comply with the AC here but is necessary is
deactivating active events because their dispatch time is shifted to
sometime earlier than the original dispatch time. If this time is still
in the future, I believe the event should still be kept as one which may
be triggered. Events which outrightly move into the past because of a
reschedule are deactivated if they aren't already completed.