Invert the Custom Card Configuration setting
We initially assumed that the default behavior for Custom Cards would be
that all Roles would have read/write access.  So, the code set all Roles,
but allowed for defining an array or Roles to exclude.  (blacklisting)

This assumption turned out to be incorrect.  Most Cards will always have
a subset of Role permissions applied.  So, this PR flips the behavior
from a blacklist to a whitelist.

By default, no permissions will be assigned except for the ones defined
in the configuration `#view_role_names` and `#edit_role_names`.

In addition, you can specify the symbol of `:all` if you happen to want
all Roles to be assigned to the Card (as we do with some sample Cards).