APERTA-9687 fixed up the frontend
The users now see the changes reflect on the same page after they change
the due date. These changes to the scheduled events take some
half-a-second on my end to reflect. Changes may be longer based on the
user's internet connection. It may be nice to ad some indicator that new
scheduled events are being loaded in the mean time.

I used flexbox to implement a faux grid. This I think may break when
there's a scheduled event with a ridiculously long name. I tried using
bootstrap, but the same issue remains because the column width has to be
fixed. An extra long scheduled-event name will force a wrap. The up
sides of this is that all the check marks are aligned nevertheless.

Also, there is a conditonal return in the ReviewReportsController: a 200
when we change the due date, and a 204 when we do not. The 200 is
necessary to reload the scheduled events which have been rescheduled.