APERTA-10968 Do not grab a bunch of permissions we should not grab
Modify the ObjectsPermissibleByRequiredPermissionsQuery class to prevent it from
overly expanding the set of permissions available to a user.

Before this change, this class would take a set of permissions that applied to a
given thing and expand it by joining with roles and then permissions. This would
yield too many permissions, permissions that should not have applied to a thing,
either because of a state restriction or a permission filter.

This may cause some issues in the front end, but it is necessary for security
purposes.