Introduce generator for creating CustomCard::Configuration class
This is useful for easing the pain of creating a valid CustomCard::Configuration
when migrating existing Tasks to custom Cards or creating a brand new custom Card.

The generator queries the local data to pull over any permissions for the existing
legacy Task that is being replaced.

Usage Example:
  rails generate custom_card:configuration "My New Cover Letter" TahiStandardTasks::CoverLetterTask

This will create:
  lib/custom_card/configurations/my_new_cover_letter.rb
  Configuration file will contain the proper permissions from the legacy task.