Delete the custom card migrator rake task
This rake task was originally created to follow the convention of the
data migrations that have existed elsewhere in aperta.  Namely, that
when a migration happens, it calls a rake task.  

Two problems with this:

1.  This rake task was really only a layer of indirection to call a 
service class which made its value questionable.

2.  Calling the same rake task more than once (even with different
arguments) in the same `db:migrate` transaction, will only ever
execute the first one.  This is due to the caching mechanism of Rake.
Yes, code could be introduced to "fix" this, but it hardly seems
worth the effort or complexity given #1 above.

Also, change the past migration for competing interests to use the
service class directly.  The code in this migration will be the proposed
way of doing task -> card migrations going forward.

Finally, make the migrator service class accept keyword arguments to
improve visibility as to what is happening.