APERTA-10881 Discussion Participant is an MS role
Previously, the Discussion Participant did not participate in Paper nor
Task. This made it a journal-level role. This scoping makes sure it's
properly scoped as a manuscript role, and makes it easier to recognize
journal-level roles.

This should not change anything with the permissions since permissions
are checked against the assignments table. All permissions in the
Discussion Participant role already apply to a Discussion Topic.