Delete old answers after they are deselected
We currently have two components which can variably display a tree of
card content with answers.

With the `if-component`, there are two alternate branches that will be
displayed. When the user selects one of these branches, the inactive
tree of answers should be deleted. With the `display-with-value`
component there is only one branch, but the answers should be similarly
deleted when the children are hidden.

In either case, these stale answer records were creating misleading
snapshots. Since the answer was never actually removed, the snapshot
never noticed any changes. With these changes, the answers will be
removed and the snapshots will now pick up the differences in answers.