APERTA-9688 Fire events when the time comes
This commit enables emails to be sent to the reviewers when the time for
the reminder comes.

To note: I used APERTA-10712 in this commit. Currently, this file is not
merged into master so I had to merge the branch here.

Two new states have been introduced into the ScheduledEvents to allow
for a time when the email fails to send.