Add notify and clean up notifyException for Bugsnag
For `notifyException`, the parameters have been reduced to just the
exception object and optional metadata. The `name` and `severity`
arguments have been removed.

A `notify` method is added to the service. It is used in places where
we want a warning, but where there may not be an exception

* Pusher failures call `notify` instead
* File upload errors use `notifyException`
* Remove 'Uncaught Ember Error'; rely on exception data
* Add file upload metadata in case of failure

APERTA-9819