APERTA-10397 (unrelated) don't pre-wrap text p whitespace
For the time being if we enter html text via CDATA as the <text></text>
content of a <content content-type="text"> element, any extra spacing in
the CDATA will be included in the rendered html.

For example
```xml
            <text>
            <![CDATA[
              <h4>Please Upload Your Source
              File</h4>
              <span>Because
              you uploaded a
              PDF, you must
              also provide
              your source
              file (e.g.
                  .tex, .docx)
              before marking
              this task as
              done.</span>
            ]]>
            </text>
```
Will be squashed into the middle of the page, since it has leading and
trailing whitespace on both sides.  This seems bad.  For the time being
I'm removing the `whitespace: pre-wrap` rule.