Move card content migration earlier, reset cache
This moves the migration to add an error message column to card_contents
earlier in the sequence of the next release, to prevent conflicts with
card loading migrations that come later.

It also clears the schema and column info caches, which we should do
for card_contents until all the card migrations are completed.