Fix Competing Interests before we break it
The 1.49 release branch contains a change to how radio buttons are
handled in custom cards. A side effect of this was to hide the header
text on the Competing Interests hard, leaving the UI with just a yes/no
radio button and nothing else.

This reverts radio buttons to the old behavior.