Dont serialize card content or xml for available_cards route
When a user hits 'add new card' on a workflow, the dialog that displays
the list of cards is currently slow.  When we fetch 'availableCards' for
a paper, those cards are serialized with all of their content and xml,
even though we never access that information.

This change only affects the available_cards route at this point.  All
other behavior remains the same (include_content defaults to true)

Speed difference for 'available_cards' route on my machine:
Before
~4seconds
After
~146ms