APERTA-11212 fix bindings for paper-file-uploader
The paper-file-uploader wasn't set up to start with a nonexistent file.
Since uploading an attachment doesn't automatically refresh the
'paper.file' or 'paper.sourcefile' it has to be done manually once a new
attachment is created.

The fileType attribute for a given attachment isn't always set by the
time the attachment is event-streamed, so it has to be computed from the
filepath as a fallback.