APERTA-11267 hook up tinymce on init
Before this change, if the rich-text-editor was initialized with 'disabled=false'
then it would never allow tinymce to save its content.  In most cases disabled
is calculated from a permission fetched asynchronously from the server - it
starts out true, then once the ajax request for the permission completes it
toggles back to false and runs the observer.  There are some cases (like if the
rich-text-editor is rendered inside an 'if' block) where the editor may
initially be rendered with 'disabled=false' and never trigger the observer.  For
things like this we add .on('init') to the observer to make sure it fires right
when the component is initialized. See
https://guides.emberjs.com/v2.15.0/object-model/observers/#toc_observers-and-object-initialization
for more detail