APERTA-10946 checked for blank fields in Liquid
This checks if any of the variables are false in the liquid template
provided. There is a false-positive edge-case where if the
`Liquid::Variable` is used in the `Liquid::Block`'s condition as a check
to render (or not) the said `Liquid::Variable`. This edge-case is
out-of-scope for this ticket, but it would be nice to note down for
later.