Disallow SUBCLASSME card titles on Choose New Card
While we are converting legacy cards to custom cards, cards with the
titles “SUBCLASSME” and “Custom Card” will continue to sneak into the
JournalTaskTypes table between the time that a card is converted to a
custom card, and the time when the legacy ruby class is actually
deleted. Trying to add either of those card types to a workflow or paper
will result in undefined behavior or an error.

This filters out cards with those titles from the choose-new-card Ember
component, with a TODO of removing card title filtering once all legacy
cards have been successfully converted to custom cards.

APERTA-11268