APERTA-11319 make font size workarounds specific
This is somewhat a temporary band aid.  Right now we have 2 ways in
place for making the question text bigger to mimic the existing cards
(I at least had a had in both !).

1. We re-use the existing '.question-text' class from '_question.scss'
This class adds a 1.4em font-size among other effects
2. We took an alternate route where we set a 1.4em font-size to children
of 'li.question .card-content-view-text p' elements in 'text-view.scss'.

These two rules interact in an unfortunate way. '1.4em' will stack
between rules so we have existing cases where a .question-text is a
child of a 'li.question .card-content-view-text p', and the font becomes
very large.

My workarounds:
1. Change the li.question css rule to only work on direct descendents.
This is the actual use case the rule is trying to address.  Since the
existing rule applies to *all* descendants it's troublesome as we
recursively render card content.
2. We also don't want .question-text to increase the font size of items
that are already affected by workaround #1, so I added a rule to null
out the font size change in a very specific fashion.

Going forward:
We shouldn't use the .question-text class in card config stuff at all.
It has too many side effects that are hard to reason about, and it ties
us to the existing styling behavior.  We want a clean slate.

Why not change the exiting custom cards to remove the class before the
release?
Because it's way way more work than a css change.