APERTA-11234 clear errors on file upload
This is the first instance where something in the card content tree needs to
communicate with the task component.  Errors are handled by the
upload-manuscript-task component itself, but we need to clear them when a user
uploads a file.

The card-event service is a bare bones wrapper around Ember.Evented.  The
service exports a helpful mixin that removes the boilerplate around subscribing
to events (if we forget to unsubscribe we'll leak memory).

I we'll be able to use the card-event service in other cases where trying to
thread actions down the card content tree is difficult or impossible.