APERTA-11196 allowed for unauthenticated routes
Before now, in Tahi-Ember, we always looked for a journal for every
Ember route. But we're meant to do this only for authenticated requests.
We have an [authorized
route](https://github.com/Tahi-project/tahi/blob/master/client/app/routes/authorized.js),
but this serves, from my knowledge, as a reactor to authorizations.

With this change, you can have unauthenticated routes run without
looking for a journal which wouldn't exist because no one is logged in.