APERTA-10397 always show file uploaders
The uploader for the manusript file was nested under a 'taskStateToggleable' if
block, which aliases to 'isEditableDueToPermissions'.

That meant that users who didn't have permission to edit the upload manuscript
task (like billing or AE) couldn't see the uploader at all.

This commit unnests the uploader from the 'if' block completely.