Don’t transition to index if you’re already there
The transitionToPaperIndex action calls code specific to the versions
component from the main paper control bar, without a check that one was
currently in versions view.

This adds a conditional to calling transitionToPaperIndex, and an Ember
test to toggle Collaborators on and off from the index view, and make
sure an error isn’t thrown.

APERTA-11301