APERTA-11218 fix polymorphic type serialization
When we serialize the type info for a relationship ember data like {id:
1, type: 'foo'} ember data expects 'foo' to be lowercase and dasherized,
matching the way it looks up internal model types in the store.  We were
sending back 'AdHocTask' instead of 'ad-hoc-task', throwing things
off.

For the curious, I followed the code path in ember's store from
store._push eventually down to store._pushResourceIdentifier, where the
store tries to find the type map for the model in question.