Eagerly create required Answer for each Repetition
Card content validations expect that Answers for any requiredField
content are created ahead of time.

Before this commit, we eagerly created any required Answers on the
server when the Task is added to the Paper. This would be Difficult™ to
create now that you can arbitrarily nest Repetitions inside each other
with different minimum numbers of repetitions.

Luckily, we have all the information we need to make this decision
immediately available to us on the client side. This allows us to save
the Answer record the first time the user opens the card.

Unfortunately, we also need to initially hide the error messages that
are created so that the user has a chance to fill out the form before
bombarding them with errors.

This should also fix a similar problem with the `if` and
`display-with-value` components. Task-creation is too early to make the
decision about whether or not the Answer should be created for a given
branch.