APERTA-11821 Actually, let's just fetch tinyMCE from the CDN for now
. . . due to performance reasons. The CDN has a prebuilt js file that contains all the plugins
we need. If we try to load the tinymce.min.js from the npm package, it doesn't contain all the
plugins, so it needs to fetch them all one-by-one after the fact, which is slow. Throwing all this
in vendor.js solved that problem, but unfortunately that's not working with IE for unknown reasons.