Destroyed Repetitions should not be returned
This is fixing a bug with the original implementation of Repetitions.
Sending back a destroyed record to Ember will keep it in the Ember
Data Store and then if it is modified on the client, it will throw
a nasty error.

In addition, returning positions of repetitions when the client is
destroying many of them can result in Ember 'inflight' errors, so
return head no content in those cases.